Benchmark for Rocket backend
